[ONNX] Create an optimize api in ONNXProgram
Move optimization from the export call to the `optimize()` method in ONNXProgram.

Users can call `optimize()` before calling `save()` to save the model. The change is made to allow users to control optimization behavior. It is not always applied in torch.onnx.export so the core export logic can be more robust.

ghstack-source-id: 17ddf6a8c6e3b66f34d55768901a63d4fced127a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/137667